mainloop: cast pointer to usize so it works on any platform

+++ src/mainloop.rs
@@ -34,7 +34,7 @@ impl MainLoop{
 	pub fn spawn(&mut self){
 		if !self.running {
 			self.running = true;
-			let gst_loop: u64 = unsafe{ mem::transmute(self.gst_loop) };
+			let gst_loop: usize = unsafe{ mem::transmute(self.gst_loop) };
 			thread::spawn( move|| {
 				unsafe{
 					g_main_loop_run ( mem::transmute(gst_loop) );
